Remove inclusion of fmpz.h in fmpz_mat.h
And deinline a couple of functions in fmpz_mat.h. Moreover, deprecate
fmpz_mat_[col/row]_equal in favor for fmpz_mat_equal_[col/row].
